test: Reorganize platform tests by module
